Expose Config object to Rails templates via #config method for template-specific configuration
